+++ scripts/unprompted.py
@@ -254,7 +254,9 @@ class Scripts(scripts.Script):
 								elif (block_name == "number"):
 									obj = gr.Number(label=this_label,value=int(content),interactive=True,info=_info,minimum=kwargs["_minimum"] if "_minimum" in kwargs else None,maximum=kwargs["_maximum"] if "_maximum" in kwargs else None)
 
-								elif (block_name == "dropdown"): obj = gr.Dropdown(label=this_label,value=content,choices=kwargs["_choices"].split(Unprompted.Config.syntax.delimiter),info=_info)
+								elif (block_name == "dropdown"):
+									_choices = Unprompted.parse_advanced(kwargs["_choices"],"wizard").split(Unprompted.Config.syntax.delimiter)
+									obj = gr.Dropdown(label=this_label,value=content,choices=_choices,info=_info)
 								elif (block_name == "radio"): obj = gr.Radio(label=this_label,choices=kwargs["_choices"].split(Unprompted.Config.syntax.delimiter),interactive=True,value=content)
 								elif (block_name == "slider"):
 									obj = gr.Slider(label=this_label,value=int(content),minimum=kwargs["_minimum"] if "_minimum" in kwargs else 1,maximum=kwargs["_maximum"] if "_maximum" in kwargs else 10,step=kwargs["_step"] if "_step" in kwargs else 1,info=_info)
@@ -478,24 +480,8 @@ class Scripts(scripts.Script):
 				info = sd_models.get_closet_checkpoint_match(Unprompted.shortcode_user_vars["sd_model"])
 				if (info): sd_models.load_model(info,None,None)
 			# control controlnet
-			elif att.startswith("controlnet"):
-				Unprompted.log(f"Setting ControlNet value: {att}")
-				try:
-					import importlib
-					cnet = importlib.import_module("extensions.sd-webui-controlnet.scripts.external_code", "external_code")
-					all_units = cnet.get_all_units_in_processing(p)
-					att_split = att.split("_") # e.g. controlnet_0_enabled
-					if len(att_split) >= 3:
-						if att_split[2] == "image":
-							from pil import Image
-							this_val = Image.open(Unprompted.shortcode_user_vars[att])
-						else: 
-							this_val = Unprompted.shortcode_user_vars[att]
-						setattr(all_units[int(att_split[1])],"_".join(att_split[2:]),this_val)
-						cnet.update_cn_script_in_processing(p, all_units)
-
-				except Exception as e:
-					Unprompted.log(f"Could not set ControlNet value: {e}",context="ERROR")
+			elif att.startswith("controlnet_") or att.startswith("cn_"):
+				Unprompted.update_controlnet_var(p,att)
 
 		if p.seed is not None and p.seed != -1.0:
 			if (Unprompted.is_int(p.seed)): p.seed = int(p.seed)